requirements.txt
flask
flask-cors
m3u8
requests
psutil
schedule
certifi
ffmpeg-python
pycryptodome
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0